Move classify docs to utilities module
